% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getApiCall}
\alias{getApiCall}
\title{Call as Soccerlab api with credentials (sessionId)}
\usage{
getApiCall(url, sessionId, raw = FALSE)
}
\arguments{
\item{url}{`string` The api url to reach}

\item{sessionId}{`string` The sessionId (available in session$userData$sessionId post login)}

\item{raw}{`bool` defaults to FALSE (TRUE = non parsed json, FALSE = default httr content parsed)}
}
\description{
Call as Soccerlab api with credentials (sessionId)
}
