% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{getSettings}
\alias{getSettings}
\title{Calculates the settings based on the config file}
\usage{
getSettings(config_file, roles = NULL)
}
\arguments{
\item{config_file}{`string` Path to the config file}

\item{roles}{`atomic vector (character)` Optional atomic (character) vector of allowed roles (UUID's)}
}
\value{
a `list` object representing the calculated settings based on config file
}
\description{
Calculates the settings based on the config file
}
