% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{getSlabToken}
\alias{getSlabToken}
\title{send message to UI to fetch token.}
\usage{
getSlabToken(session)
}
\arguments{
\item{session}{Rshiny session}
}
\description{
This function is not intended to be used by end user.
}
