% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{setupJS}
\alias{setupJS}
\title{Setup SLab JS requirements}
\usage{
setupJS(settings)
}
\arguments{
\item{settings}{`list` object to holding settings details}
}
\description{
This function is not intended to be used by end user.
}
