% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinySlabApp}
\alias{shinySlabApp}
\title{Create a Shiny app object with Soccerlab Authentication}
\usage{
shinySlabApp(
  ui,
  server,
  config_file = NULL,
  roles = NULL,
  verboseFlag = getOption("slab_verbose"),
  ...
)
}
\arguments{
\item{ui}{an ordinary UI object to create shiny apps.}

\item{server}{an ordinary server object to create shiny apps.}

\item{config_file}{path to JSON configuration file.}

\item{roles}{atomic vector (character) of allowed roles, default NULL (= allow all).}

\item{verboseFlag}{TRUE/FALSE will show all debug information (prints) AND verbose UI.}

\item{...}{Other arguments as passed on to [shiny::shinyApp()].}
}
\description{
This function modifies ui and server objects to run using Soccerlab
authentication.
}
\details{
You can also use a different configuration file by setting the
`slab_config_file` option with:
`options(slab_config_file = "path/to/your_file.json")`.
}
\section{Disable Soccerlab while developing apps}{


In order to develop without using SoccerLab authentication,
one can run options(slab_disable = TRUE) to
  disable soccerlab temporarily.
}

