% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{slab_find_config_file}
\alias{slab_find_config_file}
\title{Find the configuration file.}
\usage{
slab_find_config_file()
}
\value{
Character vector of length one contaning the path of the
  `soccerlab.json` file.
}
\description{
Tries to find the path to the `soccerlab.json` file. First, it tries to get
  this info from `options(slab_config_file = )`. If this option is `NULL`
  (the default) it tries to find the `soccerlab.json` within the working
  directory. If the file does not exist, it raises an error.
}
\seealso{
[`use_soccerlab`].
}
