% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{slab_ui}
\alias{slab_ui}
\title{Injects ui objects to facilitate in token based authentication with Soccerlab.}
\usage{
slab_ui(settings)
}
\arguments{
\item{settings}{`list` object to holding settings details}
}
\description{
This function is not intended to be used by end user.
}
