% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{use_soccerlab}
\alias{use_soccerlab}
\title{Soccerlab configuration file}
\usage{
use_soccerlab(path = ".", file = "soccerlab.json", overwrite = FALSE)
}
\arguments{
\item{path}{Directory name. Should be the root of the shiny app}

\item{file}{File name. Defaults to `soccerlab.json`.}

\item{overwrite}{Will only overwrite existing path if `TRUE`.}
}
\description{
Generates a minimal config file for soccerlab
}
