% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{verifyToken}
\alias{verifyToken}
\title{Verifies a Soccerlab issued token}
\usage{
verifyToken(token, settings)
}
\arguments{
\item{token}{`string` The token to verify}

\item{settings}{`list` The calculated settings object}
}
\value{
a `list` object containing authenticated session and user profile
}
\description{
Verifies a Soccerlab issued token
}
